+++ CHANGELOG.md
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `charCountText` property on `TextField` and `TextArea`
 - `multiple` option for `Autocomplete`
 - A warning is logged for a wrong or unknown `Icon`
+- `disabled` prop for `Button`
 
 ## [7.0.0-beta.0] - 2023-03-30